+++ server/database/init-database.ts
@@ -3,6 +3,10 @@ import { Database } from "./general/database";
 import { InMemoryDatabase } from "./in-memory/in-memory-database";
 import { MongoDatabase } from "./mongo/mongo-database";
 
+/**
+ * Ideally returns a MongoDatabase, but falls back to an InMemoryDatabase if
+ * env.DATABASE_URL is not set.
+ */
 export async function initDatabase(): Promise<Database> {
   if (env.DATABASE_URL != null) {
     return await MongoDatabase.init(env.DATABASE_URL);
